Fix Warnings and remove ownership commets (This is handled by git history.)

+++ src/common.c
@@ -1,7 +1,3 @@
-//
-// Created by Zikai Liu on 3/20/22.
-//
-
 #include "common.h"
 #include <stdio.h>
 
@@ -13,5 +9,5 @@ void print_triangle_printf(index_t a, index_t b, index_t c) {
     if (a > b) SWAP(a, b);
     if (b > c) SWAP(b, c);
 #undef SWAP
-    printf(INDEX_FMT " " INDEX_FMT " " INDEX_FMT "\n", a, b, c);
+    printf(INDEX_FMT " " INDEX_FMT " " INDEX_FMT "\n", (uint64_t) a, (uint64_t) b, (uint64_t) c);
 }
